Starting to move core build dungeon methods to their own module
