+++ DungeonProject/DungeonProject/DungeonBuilder.py
@@ -1,5 +1,7 @@
 import sys
 import random
+import Utilities
+from Utilities import *
 
 class DungeonBuilder:
 
@@ -133,6 +135,119 @@ class DungeonBuilder:
             if c != "_":
                 g_ += c
         return g_
+
+    # Build a dungeon according to the given branch from the given transform point
+    # Builds a path followed by the first room in the graph, the recursively does so for the continuing branches
+    @staticmethod
+    def buildDungeon(graph, transform, manager, kitScene, scene, placedTiles, difficultyLevel):
+        if len(graph) == 0:
+            return
+
+        #If the spawn room is to be built, don't build a path first
+        path = [transform]
+        if graph[0] != "O":
+            scene.GetRootNode().AddChild(makeBox(32, 128, 32, manager))
+            path = DungeonBuilder.buildPath(transform, manager, kitScene, scene, placedTiles, difficultyLevel)
+
+        #Obtain the branches after the room to be built
+        graphs = Utilities.split(graph[1:]) if len(graph) > 1 else [] 
+
+        #Build the room
+        properties = {
+            "isSpawnRoom": graph[0] == "O",
+            "numExits": len(graphs)
+            }
+        transform = buildRoom(properties, path[-1], manager, kitScene, scene, placedTiles)
+
+        #If failed to place room, retry
+        if not transform:
+            for back in range(len(path)-1): # Backtrack until no more path is available
+                #Remove last tile
+                node = scene.GetRootNode().GetChild(scene.GetRootNode().GetChildCount() - 1)
+                scene.GetRootNode().RemoveChild(node)
+                node.Destroy()
+                path.pop()
+                placedTiles.pop()
+
+                #Retry without adding anything
+                transform = buildRoom(properties, path[-1], manager, kitScene, scene, placedTiles)
+                #Break when successfully placed the room
+                if transform:
+                    break
+
+            #If still failed to place the room, stop this branch's generation
+            if not transform:
+                return
+
+        #Recursively build the next part of the dungeon
+        for i in range(len(graphs)):
+            buildDungeon(graphs[i], transform[i], manager, kitScene, scene, placedTiles, difficultyLevel)
+
+    # Builds a path from the given transform point
+    # Returns the sequence of path transforms (allowing for simple backtracking)
+    # Path transform: end point of the path, from where to build the next part of the dungeon
+    # For now, paths should only return one path, since they are built between one room and another
+    @staticmethod
+    def buildPath(transform, manager, kitScene, scene, placedTiles, difficultyLevel):
+        transforms = [transform]
+        weights = []
+
+        # Choose the length of the corridor depending on the chosen difficulty
+        numTiles = random.randint(*[[5, 10], [5, 15], [10, 20], [15, 25], [20, 30]][difficultyLevel-1])
+
+        # Repeat the generation until the number of tiles to place has been reached
+        i = 0
+        while i < numTiles: 
+            ret = False
+            # If the generator comes from backtracking, use the remaining weights that 
+            #   were not explored when generating this tile
+            if len(weights) <= i:
+                weights.append({
+                        400: [
+                            {0: 10, 1: 1, 2: 1, 8: 1},
+                            {0: 10, 1: 2.5, 2: 2.5, 8: 1},
+                            {0: 10, 1: 2.5, 2: 2.5, 10: 1, 11: 1, 8: 1, 22: 0.5},
+                            {0: 10, 1: 3.33, 2: 3.33, 10: 2.5, 11: 2.5, 8: 1, 22: 1},
+                            {0: 10, 1: 5, 2: 5, 10: 3.33, 11: 3.33, 8: 1, 22: 1.5}
+                        ],
+                        800: [
+                            {12: 10, 13: 1, 14: 1, 24: 1},
+                            {12: 10, 13: 2.5, 14: 2.5, 24: 1},
+                            {12: 10, 13: 2.5, 14: 2.5, 28: 1, 29: 1, 24: 1, 23: 0.25, 40: 0.25},
+                            {12: 10, 13: 3.33, 14: 3.33, 28: 2.5, 29: 2.5, 24: 1, 23: 0.25, 40: 0.25},
+                            {12: 10, 13: 5, 14: 5, 28: 3.33, 29: 3.33, 24: 1, 23: 0.25, 40: 0.25}
+                        ],
+                        1600: [
+                            {30: 10, 31: 1, 32: 1, 42: 1},
+                            {30: 10, 31: 2.5, 32: 2.5, 42: 1},
+                            {30: 10, 31: 2.5, 32: 2.5, 42: 1, 41: 2.5},
+                            {30: 10, 31: 3.33, 32: 3.33, 42: 1, 41: 6},
+                            {30: 10, 31: 5, 32: 5, 42: 1, 41: 8.16}
+                        ],
+                    }[transforms[-1][4]][difficultyLevel-1])
+
+            # Try with all alternatives
+            while not ret and len(weights[-1]) > 0:
+                tile = randomWeightedChoice(weights[-1])
+                ret = TileGenerator.generateTile(tile, transforms[-1], manager, kitScene, scene, placedTiles)
+                del weights[-1][tile]
+
+            if ret:
+                transforms += [ret[0]]
+                i += 1
+            # If all alternatives failed, backtrack and continue generation
+            else: 
+                if i == 0: # Do not backtrack if no more elements available
+                    break
+                node = scene.GetRootNode().GetChild(scene.GetRootNode().GetChildCount() - 1)
+                scene.GetRootNode().RemoveChild(node)
+                node.Destroy()
+                placedTiles.pop()
+                weights.pop()
+                transforms.pop()
+                i -= 1
+
+        return transforms
